Compiler server working
Added specs, have enough to move to front end now and work on that.
Currently accepts programs a line at a time through the socket, responds on each line. Tests include getting imported
data, getting code back.
This is NOT a full round trip test as that involves the file system and so is changeable with each test..