Revised compiler server messages for files
Revised handling of these, which now means that we include state (last status) and a map of file names to file contents for all imports. Interestingly, no changes were needed in the actual compiler. We only needed to add a fail message in the file provider saying which import file was not yet added.

Compiles are attempted now as soon as a qplprogram is read in, and after each import file is read. This will either respond with a success message including the QPO code, a failure with message in case of a syntax or other error, or a "need file" message. The need file is the signal to try an import.

Areas to firm up (not necessarily exhaustive :)
  - Handling of "file not found" for imports - may currently loop (short term)
  - Perhaps add "pre-analysis" of imports and ask up front. (long term - also note that this would require changes in the server as it currently resets the "state" every time a new quantum program is read.