Add control panel to allow tests to pass
designed main screen to not show at start - become visible once program is loaded. Program is
now at the point where the emulator server must be started for tests to run.
Once a program is loaded, the control panel is set to visible, from there we can add button presses and so on.