Setting up build tasks
Concentrated on making RAKE work today. Correctly builds and runs

Need to change default on IntelliJ forms - have them generate to the java source files, otherwise this is dependant upon the intellij compiler. Sigh.
May have been a mistake to go with the forms designer. May remove that entirely in a future version.