Fixed issue #1 - type unification issue
Added  a test to prove out the problem - "Typevariabla 'a'" should unify with "INT" when part of a call. See reverse.qpl for example, failed on the call to reverse.

Issue was that unification did not allow this for anything but malleable variables - however, in this call, a "TypeVariable" is malleable, so just added an extra
check for that.