Significant FEST progress
And some refactoring…

Main issue remaining is that the SWING app will NOT shut down in between scenarios. It seems to be intimately tied in with the cucumber JVM. For example, If I actually do shut it down via using the robot to do a command-Q, it stops cucumber as well.

Because it is still running, the Dump and other panels are created anew with each scenario, then when searching for them, the cucumber system finds multiple instances - and complains.

With Jemmy, I had just left the app running from scenario to scenario - not even trying to shut it down or restart the system. This, however, is somewhat anathema to the concept of clean-room testing, as is supposed to be the case with Cucumber.

Need to consider how best to address this.

Note that when running, a close of the main screen does shut down everything (added disposes of all dialogs and panels - did NOT call System.Exit).