Fixed issue with simplecov not working on cucumber. Now at 99% test coverage.
Need to add a couple of tests regarding depth and call multiplier, but discovered the emulator is not working quite as expected. Even when I reset the default call depth to 1 in the haskell code, the emulator seems to perform at least 10 calls. Need to track this down and then write a couple of cucumber tests to spin those elements.