Updated for GHC7.8.3
Primarily added instances of Applicative (when instance of Monad) and
Alternative (when instance of MonadPlus).

Bigger thing though is we now need to use the sandbox method for
developing. Doing a base install of dependancies in "public space" leads
to warnings about reinstalling text and unordered-containers and then leads
mismatches when trying to compile.

Issue seems to be Aeson defaulting to 8 when doing base install, but does
7.0.6 when using a sandbox. Hmmmm